Refactor : 검색 페이지 build 오류 수정
검색 페이지에서 검색클라이언트 구성 요소 사용
